requirements.txt
asgiref==3.9.1
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
sqlparse==0.5.3
gunicorn
rich
